requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pillow>=10.4.0
exifread>=3.0.0
reportlab>=4.2.0
